Move monitor into child process

+++ launch_service_c/src/bag_service_c.cpp
@@ -57,7 +57,7 @@ int main(int argc, char **argv)
 
         // Create extra launch arguments
         std::vector<std::string> launch_args;
-        for (size_t i = child_flag_index + 1; i < argc; ++i)
+        for (int i = child_flag_index + 1; i < argc; ++i)
         {
             launch_args.push_back(argv[i]);
         }
